requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
black==21.6b0
chardet==4.0.0
click==8.0.1
discord.py==1.7.3
idna==3.2
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
python-dotenv==0.18.0
regex==2021.7.6
toml==0.10.2
typing-extensions==3.10.0.0
yarl==1.6.3
